Changes for LAZY. Removed etext stuff from stack.c. Removed lazyref_gc.c
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@887 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/GC/aoatoioa.c
@@ -72,7 +72,7 @@ void AOAtoIOAInsert( theCell)
 #ifdef RTLAZY
     if ((*(int *) theCell) < 0) {
       /* This could happen if called from extobjinterface.assignRef. */
-      negAOArefsINSERT (theCell);
+      negAOArefsINSERT ((long) theCell);
       return;
     }
 #endif
